+++ modules/engine/test/quirky-json-test.js
@@ -1,0 +1,67 @@
+/*
+ * Copyright 2012 eBay Software Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+var Engine = require('../lib/engine'),
+    http = require('http'),
+    util = require('util');
+
+//
+// There are many conventions to convert xml to json. If the default xml2json does not work you,
+// you can override it on table by table basis.
+//
+// To override, specify path to an xformers.json file. Here is an example.
+//
+// {
+//   "some.table" : "modulename"
+// }
+//
+// The module is loaded related to process.cwd()
+//
+module.exports = {
+    'quirky json': function (test) {
+        var server = http.createServer(function (req, res) {
+            res.writeHead(200, {
+                'Content-Type': 'application/xml'
+            });
+            res.write('<hello>world</hello>');
+            res.end();
+        });
+
+        server.listen(3000, function () {
+            // Do the test here.
+            var engine = new Engine({
+                tables: __dirname + '/quirky-json',
+                xformers: __dirname + '/quirky-json/xformers.json'
+            });
+            engine.execute('create table quirky.json on select get from "http://localhost:3000"; return select * from quirky.json',
+                function (emitter) {
+                emitter.on('end', function (err, result) {
+                    if(err) {
+                        console.log(err.stack || util.inspect(err, false, 10));
+                        test.fail('got error');
+                        test.done();
+                    }
+                    else {
+                        test.equals(result.headers['content-type'], 'application/json', 'json expected');
+                        test.deepEqual(result.body, { "HELLO" : "WORLD"});
+                        test.done();
+                    }
+                    server.close();
+                });
+            });
+        });
+    }
+};