+++ modules/engine/test/quirky-json/quirky.xml2json.js
@@ -1,0 +1,6 @@
+
+exports.toJson = function(data, respCb, errorCb) {
+    return respCb({
+        "HELLO" : "WORLD"
+    });
+};